Rename eval to evaluation
Reason: eval is already a python keyword
